+++ app/models/knowledge_base/category_article_association.rb
@@ -1,6 +1,5 @@
 module KnowledgeBase
   class CategoryArticleAssociation < ActiveRecord::Base
-    belongs_to :category
-    belongs_to :article
+    include Concerns::Models::CategoryArticleAssociation
   end
 end
