+++ app/models/knowledge_base/sectionables/attachments/attachment.rb
@@ -1,9 +1,5 @@
-module KnowledgeBase::Sectionables::Attachments
-  class Attachment < ActiveRecord::Base
-    belongs_to :list, foreign_key: :attachment_list_id
-
-    mount_uploader :file, KnowledgeBase::GenericFileUploader
-
-    validates :title, :file, :list, presence: true
+module KnowledgeBase
+  class Sectionables::Attachments::Attachment < ActiveRecord::Base
+    include Concerns::Models::Sectionables::Attachments::Attachment
   end
 end
