+++ app/models/knowledge_base/sectionables/attachments/list.rb
@@ -1,13 +1,5 @@
-module KnowledgeBase::Sectionables::Attachments
-  class List < ::KnowledgeBase::Sectionables::Sectionable
-    has_many :sections, as: :sectionable, dependent: :destroy
-
-    has_many :attachments, dependent: :destroy, inverse_of: :list,
-      foreign_key: :attachment_list_id
-    accepts_nested_attributes_for :attachments, allow_destroy: true
-
-    def to_s
-      "#{super} ##{id}: #{title}"
-    end
+module KnowledgeBase
+  class Sectionables::Attachments::List < Sectionables::Sectionable
+    include Concerns::Models::Sectionables::Attachments::List
   end
 end
