+++ app/models/knowledge_base/sectionables/gallery.rb
@@ -1,14 +1,5 @@
-module KnowledgeBase::Sectionables
-  class Gallery < Sectionable
-    has_many :images, -> { order position: :asc },
-      dependent: :destroy
-
-    has_many :sections, as: :sectionable, dependent: :destroy
-
-    accepts_nested_attributes_for :images, allow_destroy: true
-
-    def to_s
-      "#{super} ##{id}: #{title}"
-    end
+module KnowledgeBase
+  class Sectionables::Gallery < Sectionables::Sectionable
+    include Concerns::Models::Sectionables::Gallery
   end
 end
