+++ app/models/knowledge_base/sectionables/gallery/image.rb
@@ -1,7 +1,5 @@
 module KnowledgeBase
   class Sectionables::Gallery::Image < ActiveRecord::Base
-    belongs_to :gallery
-
-    mount_uploader :image, KnowledgeBase.config.gallery_image_uploader
+    include Concerns::Models::Sectionables::Gallery::Image
   end
 end
