+++ app/models/knowledge_base/sectionables/links/link.rb
@@ -1,5 +1,5 @@
-module KnowledgeBase::Sectionables::Links
-  class Link < ActiveRecord::Base
-    belongs_to :list, foreign_key: :links_id
+module KnowledgeBase
+  class Sectionables::Links::Link < ActiveRecord::Base
+    include Concerns::Models::Sectionables::Links::Link
   end
 end
