+++ app/models/knowledge_base/sectionables/links/list.rb
@@ -1,13 +1,5 @@
-module KnowledgeBase::Sectionables::Links
-  class List < ::KnowledgeBase::Sectionables::Sectionable
-    has_many :links, dependent: :destroy, foreign_key: :links_id
-
-    has_many :sections, as: :sectionable, dependent: :destroy
-
-    accepts_nested_attributes_for :links, allow_destroy: true
-
-    def to_s
-      "#{super} ##{id}: #{links.map(&:title).join(', ')}"
-    end
+module KnowledgeBase
+  class Sectionables::Links::List < Sectionable
+    include Concerns::Models::Sectionables::Links::List
   end
 end
