+++ app/models/knowledge_base/sectionables/list.rb
@@ -1,15 +1,5 @@
-module KnowledgeBase::Sectionables
-  class List < Sectionable
-    mount_uploader :image, KnowledgeBase.config.list_image_uploader
-
-    has_many :items, -> { order 'position ASC' }
-
-    has_many :sections, as: :sectionable, dependent: :destroy
-
-    accepts_nested_attributes_for :items, allow_destroy: true
-
-    def to_s
-      "#{super} ##{id}: #{title}"
-    end
+module KnowledgeBase
+  class Sectionables::List < Sectionables::Sectionable
+    include Concerns::Models::Sectionables::List
   end
 end
