+++ app/models/knowledge_base/sectionables/text.rb
@@ -1,11 +1,5 @@
-module KnowledgeBase::Sectionables
-  class Text < Sectionable
-    mount_uploader :image, KnowledgeBase.config.text_image_uploader
-
-    has_many :sections, as: :sectionable, dependent: :destroy
-
-    def to_s
-      "#{super} ##{id}: #{heading}"
-    end
+module KnowledgeBase
+  class Sectionables::Text < Sectionables::Sectionable
+    include Concerns::Models::Sectionables::Text
   end
 end
