+++ app/models/knowledge_base/sectionables/video.rb
@@ -1,13 +1,5 @@
-module KnowledgeBase::Sectionables
-  class Video < Sectionable
-    include Embeddable
-
-    embeddable :video, from: :url
-
-    has_many :sections, as: :sectionable, dependent: :destroy
-
-    def to_s
-      "#{super} ##{id}: #{url}"
-    end
+module KnowledgeBase
+  class Sectionables::Video < Sectionables::Sectionable
+    include Concerns::Models::Sectionables::Video
   end
 end
