+++ lib/knowledge_base/concerns/models/category_article_association.rb
@@ -1,0 +1,8 @@
+module KnowledgeBase::Concerns::Models::CategoryArticleAssociation
+  extend ActiveSupport::Concern
+
+  included do
+    belongs_to :category
+    belongs_to :article
+  end
+end
