+++ lib/knowledge_base/concerns/models/sectionables/links/link.rb
@@ -1,0 +1,7 @@
+module KnowledgeBase::Concerns::Models::Sectionables::Links::Link
+  extend ActiveSupport::Concern
+
+  included do
+    belongs_to :list, foreign_key: :links_id
+  end
+end
