+++ lib/knowledge_base/concerns/models/sectionables/list.rb
@@ -1,0 +1,17 @@
+module KnowledgeBase::Concerns::Models::Sectionables::List
+  extend ActiveSupport::Concern
+
+  included do
+    mount_uploader :image, KnowledgeBase.config.list_image_uploader
+
+    has_many :items, -> { order 'position ASC' }
+
+    has_many :sections, as: :sectionable, dependent: :destroy
+
+    accepts_nested_attributes_for :items, allow_destroy: true
+
+    def to_s
+      "#{super} ##{id}: #{title}"
+    end
+  end
+end
