+++ lib/knowledge_base/concerns/models/sectionables/list/item.rb
@@ -1,0 +1,7 @@
+module KnowledgeBase::Concerns::Models::Sectionables::List::Item
+  extend ActiveSupport::Concern
+
+  included do
+    belongs_to :list
+  end
+end
